pyproject.toml
[project]
name = "test-backend-task"
version = "0.1.0"
description = "test_backend_task"
authors = [
    {name = "Dima Charkin",email = "115488709+Millitarian@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

[tool.poetry.dependencies]
advanced-alchemy = ">=1.8.0,<2.0.0"
aio-pika = ">=9.5.8,<10.0.0"
aiormq = ">=6.9.2,<7.0.0"
alembic = ">=1.17.2,<2.0.0"
annotated-types = ">=0.7.0,<0.8.0"
anyio = ">=4.11.0,<5.0.0"
asyncpg = ">=0.31.0,<0.32.0"
certifi = ">=2025.11.12,<2026.0.0"
click = ">=8.3.1,<9.0.0"
colorama = ">=0.4.6,<0.5.0"
EditorConfig = ">=0.17.1,<0.18.0"
Faker = ">=38.2.0,<39.0.0"
fast-depends = ">=3.0.3,<4.0.0"
fast-query-parsers = ">=1.0.3,<2.0.0"
faststream = ">=0.6.3,<0.7.0"
granian = ">=2.6.0,<3.0.0"
greenlet = ">=3.2.4,<4.0.0"
h11 = ">=0.16.0,<0.17.0"
httpcore = ">=1.0.9,<2.0.0"
httptools = ">=0.7.1,<0.8.0"
httpx = ">=0.28.1,<0.29.0"
idna = ">=3.11,<4.0.0"
Jinja2 = ">=3.1.6,<4.0.0"
jsbeautifier = ">=1.15.4,<2.0.0"
litestar = ">=2.18.0,<3.0.0"
litestar-asyncpg = ">=0.5.0,<0.6.0"
litestar-granian = ">=0.14.2,<0.15.0"
litestar-htmx = ">=0.5.0,<0.6.0"
Mako = ">=1.3.10,<2.0.0"
markdown-it-py = ">=4.0.0,<5.0.0"
MarkupSafe = ">=3.0.3,<4.0.0"
mdurl = ">=0.1.2,<0.2.0"
msgspec = ">=0.20.0,<0.21.0"
multidict = ">=6.7.0,<7.0.0"
multipart = ">=1.3.0,<2.0.0"
pamqp = ">=3.3.0,<4.0.0"
pika = ">=1.3.2,<2.0.0"
polyfactory = ">=3.1.0,<4.0.0"
propcache = ">=0.4.1,<0.5.0"
pydantic = ">=2.12.5,<3.0.0"
pydantic-core = ">=2.41.5,<3.0.0"
Pygments = ">=2.19.2,<3.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
PyYAML = ">=6.0.3,<7.0.0"
rich = ">=14.2.0,<15.0.0"
rich-click = ">=1.9.4,<2.0.0"
setproctitle = ">=1.3.7,<2.0.0"
shellingham = ">=1.5.4,<2.0.0"
six = ">=1.17.0,<2.0.0"
sniffio = ">=1.3.1,<2.0.0"
SQLAlchemy = ">=2.0.44,<3.0.0"
structlog = ">=25.5.0,<26.0.0"
typer = ">=0.19.2,<0.20.0"
typing-inspection = ">=0.4.2,<0.5.0"
typing_extensions = ">=4.15.0,<5.0.0"
tzdata = ">=2025.2,<2026.0"
uvicorn = ">=0.38.0,<0.39.0"
watchfiles = ">=1.1.1,<2.0.0"
websockets = ">=15.0.1,<16.0.0"
yarl = ">=1.22.0,<2.0.0"
redis = "^7.1.0"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
